chore: 스크롤 수정 & 댓글 컴포넌트 분리
