+++ SMUPOOL/src/layout/BoardLayout/BoardLayout.style.js
@@ -14,8 +14,7 @@ const Wrapper = styled.div`
   flex: 1;
 
   height: 93%;
-  width: 100%;
-  overflow-x: scroll;
+  overflow-x: auto;
 `;
 
 export { Container, Wrapper };
